requirements.txt

PyPDF2==3.0.1
langchain==0.1.0
langchain-google-genai==0.0.5
google-generativeai==0.3.1
faiss-cpu==1.7.4
python-dotenv==1.0.0